requirements.txt
pyserial
gpiozero
pigpio
colorzero
digi-xbee
spidev
epydoc
mfrc522
